Fixed the DDL file, added AbstractEntity

+++ src/main/resources/db/DDL.sql
@@ -1,6 +1,6 @@
 create table user
 (
-    id int unsigned auto_increment primary key,
+    id bigint unsigned auto_increment primary key,
     email varchar(250) unique,
     password varchar(60),
     name varchar(50)
